+++ lib/src/default_index/revset_engine.rs
@@ -1161,6 +1161,14 @@ fn build_predicate_fn(
                 Ok(pattern.matches(commit.description()))
             })
         }
+        RevsetFilterPredicate::Subject(pattern) => {
+            let pattern = pattern.clone();
+            box_pure_predicate_fn(move |index, pos| {
+                let entry = index.entry_by_pos(pos);
+                let commit = store.get_commit(&entry.commit_id())?;
+                Ok(pattern.matches(commit.description().lines().next().unwrap_or_default()))
+            })
+        }
         RevsetFilterPredicate::AuthorName(pattern) => {
             let pattern = pattern.clone();
             box_pure_predicate_fn(move |index, pos| {